Rename files
DE-11289 #time 5m
