+++ lib/broadside.rb
@@ -4,8 +4,8 @@ require 'broadside/configuration'
 require 'broadside/configuration/struct'
 require 'broadside/configuration/aws'
 require 'broadside/configuration/base'
-require 'broadside/configuration/deploy'
-require 'broadside/configuration/ecs'
+require 'broadside/configuration/deploy_config'
+require 'broadside/configuration/ecs_config'
 require 'broadside/deploy'
 require 'broadside/deploy/ecs'
 require 'broadside/version'
