+++ lib/broadside/configuration/ecs_config.rb
